feat(routes): /ontology/type
