+++ frontend/src/routes/_superuser/ontology/$type.module.scss
